Добавил вертку 4 и выполнил задание.
